Upgrade pyramid_tm
Upgrade pyramid_tm to v2.x. The significant change in this version is
that pyramid_tm is now positioned _over_ the EXCVIEW tween which means
that it can be active in exception views. This is useful because some of
our exception views may access the DB to look up user info, feature
flags etc.