Ignore query parameters which do not match searchable fields in `/api/search` queries
Instead of blindly adding every query parameter not handled by a
specific filter/matcher/aggregator to a search request as a key-value
matcher in the Elasticsearch query, only add those parameters which
match names of fields in the index.

This means that generic API query parameters, general h query parameters
(eg. '__feature__[...]') or unknown query parameters will not affect the
results.

Some of these generic parameters may be consumed before they ever reach
the `h.search` APIs, but it is not always possible to rely on this. For
example `request.feature` can be called at any point in the request and
it consults `request.GET`.