Experiment with rate-limiting API requests based on IP and auth header.
Use nginx's rate limiting to rate-limit requests based on the auth token, if
provided, or the IP otherwise.

The idea is to limit requests by IP for logged-out users but allow logged-in
users to continue making requests even if their IP would otherwise be rate
limited.

IMPORTANT CAVEAT: The Authorization header is controlled by the requester, so it
is trivial to bypass nginx-level checks. However any dummy values submitted will
fail auth checks in the Pyramid app.